Hackrx_Kortex/requirements.txt
fastapi==0.109.0
uvicorn==0.24.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
aiohttp==3.9.1
python-dotenv==1.0.0
pinecone-client==2.2.4
openai==1.3.8
requests==2.31.0
pypdf2==3.0.1
docx2txt==0.8
email-validator==2.1.0.post1
numpy==1.24.3
pytest==7.3.1
pytest-asyncio==0.21.0
httpx==0.24.1
slowapi==0.1.8
psutil==5.9.5